updating parallelization

+++ internal/actions/clean.go
@@ -7,7 +7,7 @@ import (
 	"plex-go-sync/internal/logger"
 )
 
-func CleanPlaylist(playlist *Playlist, config *Config, fs filesystem.FileSystem) (map[string]uint64, uint64, string, error) {
+func CleanPlaylist(playlist *Playlist, config *Config, fs filesystem.FileSystem) (map[string]uint64, int64, string, error) {
 	var playlistItems []PlaylistItem
 	var playlistItemMap map[string]bool
 	var err error
@@ -34,7 +34,7 @@ func CleanPlaylist(playlist *Playlist, config *Config, fs filesystem.FileSystem)
 	logger.LogInfo("Cleaning ", path.Join(fs.GetPath(), base), " directory")
 	existing, existingSize, err := fs.Clean(base, playlistItemMap)
 	if err == nil {
-		logger.LogInfo("Existing Size: ", humanize.Bytes(existingSize))
+		logger.LogInfo("Existing Size: ", humanize.Bytes(uint64(existingSize)))
 	}
 
 	return existing, existingSize, base, err
